moved rebuilding the cache to a background thread
When launching the app from Xcode I recognized, that it didn't come to front immediately. This was due to the app rebuilding the cache on the main thread. As a solution I moved the cache rebuilding on a background thread but made sure, that the UI is updated on the main thread.